+++ libs/web/library/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.3](https://github.com/ever-co/ever-capture/compare/library-0.0.2...library-0.0.3) (2024-11-21)
+
+### Dependency Updates
+
+* `convert-video-feature` updated to version `0.1.2`
+* `screenshot-feature` updated to version `0.1.1`
+
+
 ## [0.0.2](https://github.com/ever-co/ever-capture/compare/library-0.0.1...library-0.0.2) (2024-11-21)
 
 ### Dependency Updates