+++ libs/web/search/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.2](https://github.com/ever-co/ever-capture/compare/search-0.0.1...search-0.0.2) (2024-11-21)
+
+### Dependency Updates
+
+* `screenshot-data-access` updated to version `0.1.1`
+
+
 ## 0.0.1 (2024-11-20)
 
 ### Dependency Updates