+++ libs/web/timesheet/data-access/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.2](https://github.com/ever-co/ever-capture/compare/timesheet-data-access-0.0.1...timesheet-data-access-0.0.2) (2024-11-21)
+
+### Dependency Updates
+
+* `notification-data-access` updated to version `0.0.2`
+* `screenshot-data-access` updated to version `0.1.1`
+* `shared-service` updated to version `0.1.1`
+
+
 ## 0.0.1 (2024-11-20)
 
 ### Dependency Updates